[error] track error line
This commit adds a optional feature which tracks where an error
occurred. This is useful when debugging an error which can be raised by
multiple places.

The feature is only enabled in "Debug" build by default.

+++ src/include/common/code_utils.hpp
@@ -93,13 +93,13 @@
  *
  * @param[in]  aStatus     A scalar status to be evaluated against zero (0).
  */
-#define SuccessOrExit(aStatus) \
-    do                         \
-    {                          \
-        if ((aStatus) != 0)    \
-        {                      \
-            goto exit;         \
-        }                      \
+#define SuccessOrExit(aStatus)              \
+    do                                      \
+    {                                       \
+        if (static_cast<int>(aStatus) != 0) \
+        {                                   \
+            goto exit;                      \
+        }                                   \
     } while (false)
 
 /**
